# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PayMyDine Restaurant API
  description: |
    Multi-tenant restaurant ordering and management system API.
    
    **Security Warning**: This API currently has no authentication implemented.
    All endpoints are publicly accessible.
    
    **Multi-tenant**: Each restaurant has isolated data through subdomain-based routing.
  version: 1.0.0
  contact:
    name: PayMyDine Support
    email: support@paymydine.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://{tenant}.paymydine.com/api/v1
    description: Production server (tenant-specific)
security:
  - {} # No authentication currently implemented
paths:
...
  /menu:
    get:
      summary: Get menu items
      description: Retrieve all available menu items with categories
      operationId: getMenu
      tags:
        - Menu
      responses:
        '200':
          description: Menu items retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/MenuItem'
                      categories:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /table-info:
    get:
      summary: Get table information
      description: Retrieve table information by table ID
      operationId: getTableInfo
      tags:
        - Tables
      parameters:
        - name: table_id
          in: query
          required: true
          description: Table ID
          schema:
            type: string
            example: "7"
      responses:
        '200':
          description: Table information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Table'
  /table-menu:
    get:
      summary: Get table menu
      description: Retrieve menu for a specific table
      operationId: getTableMenu
      tags:
        - Tables
      parameters:
        - name: table_id
          in: query
          required: true
          description: Table ID
          schema:
            type: string
            example: "7"
        - name: location_id
          in: query
          required: false
          description: Location ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Table menu retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/MenuItem'
                      categories:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'
...
components:
  schemas:
    MenuItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Margherita Pizza"
        description:
          type: string
          example: "Classic pizza with tomato and mozzarella"
        price:
          type: number
          format: float
          example: 15.99
        image:
          type: string
          example: "/images/pizza.jpg"
        category_id:
          type: integer
          example: 1
        category_name:
          type: string
          example: "Pizza"
        calories:
          type: integer
          example: 300
        allergens:
          type: array
          items:
            type: string
          example: ["gluten", "dairy"]
        stock_qty:
          type: integer
          example: 10
        minimum_qty:
          type: integer
          example: 1
        available:
          type: boolean
          example: true
    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Pizza"
        description:
          type: string
          example: "Italian pizzas"
        image:
          type: string
          example: "/images/pizza-category.jpg"
...
    Table:
      type: object
      properties:
        table_id:
          type: integer
          example: 7
        table_name:
          type: string
          example: "Table 7"
        min_capacity:
          type: integer
          example: 2
        max_capacity:
          type: integer
          example: 4
        table_status:
          type: integer
          example: 1
        qr_code:
          type: string
          example: "ms10zQpxP"
        location_id:
          type: integer
          example: 1
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Order not found"
        message:
          type: string
          example: "The requested order does not exist"
...